[FIX] survey: replace the native datetimepicker by an Odoo-styled BS one
Commit: 4e6fdd345cff4f156902b3c066aa065b7634ab96
with a starking lack of foresight, erroneously put a native input date.
It wrongly assumed that the text input was a text input instead of a BS
datetimepicker, which was broken when upgrading BS3 to BS4.
Which means that it wasn't styled as the rest as Odoo, as well as not being
configurable, not working on some browsers, and subject to locale errors.

We fix the DOM so that the BS datetimepicker is correctly applied.

Related to 9a6cd4e41dda56c4f454ed898ed17adad0d8556f

closes odoo/odoo#28135