[REF] survey: break dependency from survey composer on mail composer
Purpose of this commit is to clean survey composer model. It currently
inherits from mail.compose.message which also inherits from mail.message.
That means a lot of fields and behavior coming from both models are present
on the survey composer. Most of them are not necessary as

 * send_mail method of mail composer is overriden and super is not called.
   Nothing from the complex behavior of mail.compose.message is used;
 * templates are just basically supported in survey, notably mass mailing or
   mass posting mode are not used;
 * methods from mail.message should not be available on survey composer;
 * most m2m fields coming from mail.message and composer have no use here;

In this commit we simply add the required fields and template behavior on
the survey composer and remove inheritance on mail.compose.message.

This commit is linked to task ID 1909804 and PR #28783.