[IMP] hr_holidays: Add leave code for payslips
Purpose
=======

Before this commit, it was not straightforward to use leave records in salary
rules programatically.

This commit implements the standard and python-code compatible way of defining
leave codes which then can be used in the salary rules in the used way.

This way, the naming of leaves will not be restricted to python-attribute
compatible naming, when one wants to use this feature.

Leave names (translatable) are an unreliable identification in python code salary rules.
A language switch would break python salary rules that use a leave type as input

TaskID: 1903964
Closes #28240

Co-authored-by: blaggacao <david@elaleman.co>