[IMP] hr_holidays: add a leave request without validation mechanism
Purpose
=======

Allow to create a leave type for which leaves are not required to be approved
by a manager.

When a leave is created, it is automatically approved if the
validation type is set to "No validation".

It is useful if a manager does not need to manually approve some leave types.

Specification
=============

Add a validation type option for 'hr.leave.request': 'no_validation'

Automatically approve the leave when created and add the approver as follower
so that he can be notified about the leave creation/validation.

closes odoo/odoo#28268