[FIX] account_budget: properly calculate percentage sum in read_group
The percentage field is computed in _compute_percentage as a value
between 0 and 1. This is good because that's what widget="percentage"
in views expects.

This makes read_group also return a value between 0 and 1, instead of
0 and 100. Otherwise list views will show e.g. 5000% when grouping
lines that sum to 50%.

Introduced in saas-11.4 at odoo/odoo@7286f81ca6e6235ce1fa1cc2f34a72.

*Do not forward port, already ported to 12.0 via
 odoo/enterprise#2975 (odoo/enterprise@a1484e650cb013f35755d)*

opw-1904481