[FIX] snailmail: avoid spam on user mailbox
the fix merged in 4e8a07b09e82779dfaecbe85d7703d3edc087acf
was only tackling half the problem. When you assign a user to an
activity, the said user will be notified by email. This also occur if
you change the assignee to the next activity even if you change it to
the same user.

This commit ensures that we only write on the user_id field when we
create the next activity and not during the update.