[FIX] sale_timesheet: UOM
Revert commits b1a9eecefbabe, 099a7fcd1c1ed2 and 5e829af8c340659. They
introduce an incorrect UOM.

The original issue was an `AccessError` due to the fact that the
employee chosen was not in the same company than the user. Therefore,
instead of changing the logic of the UOM selection, we make sure to
select the employee in the appropriate company first, then fall back on
any company then.

opw-1888299